Visitors, device-detection, hubspot and dashboards
Also I moved some functions from admin to users.php and reorganize code
to each class (options admin page in options.php)
