+++ functions/utilities/utilities.php
@@ -11,70 +11,50 @@
 // Exit if accessed directly
 if( !defined( 'ABSPATH' ) ) exit;
 
-use DeviceDetector\DeviceDetector;
-
 if( !class_exists( 'Nebula_Utilities' ) ) {
 
     class Nebula_Utilities {
 
+        /**
+         * @var         Nebula_Visitors Nebula visitors
+         * @since       1.0.0
+         */
+        public $visitors;
+
+        /**
+         * @var         Nebula_Hubspot_CRM Nebula hubspot crm
+         * @since       1.0.0
+         */
+        public $hubspot_crm;
+
+        /**
+         * @var         Nebula_Device_Detection Nebula device detection
+         * @since       1.0.0
+         */
+        public $device_detection;
+
         public function __construct() {
+            // Utilities classes
+            require_once( __DIR__ . 'visitors.php');
+            require_once( __DIR__ . 'hubspot-crm.php');
+            require_once( __DIR__ . 'device-detection.php');
+
+            // TODO: Keep in this structure let us to make calls like nebula()->utilities->visitors->update_visitor();
+            // TODO: If you want use nebula()->visitors->update_visitor(); we need to move this to nebula.php
+
+            $this->visitors = new Nebula_Visitors();
+            $this->hubspot_crm = new Nebula_Hubspot_CRM();
+            $this->device_detection = new Nebula_Device_Detection();
+
             //Used to detect if plugins are active. Enables use of is_plugin_active($plugin)
             require_once(ABSPATH . 'wp-admin/includes/plugin.php');
             require_once(ABSPATH . 'wp-admin/includes/file.php');
 
+            // TODO: Move into Nebula_Google_Analytics or Nebula_Utilities_Google_Analytics (I prefer Nebula_Google_Analytics)
+
             //Sends events to Google Analytics via AJAX (used if GA is blocked via JavaScript)
             add_action('wp_ajax_nebula_ga_event_ajax', array( $this, 'nebula_ga_event_ajax' ) );
             add_action('wp_ajax_nopriv_nebula_ga_event_ajax', array( $this, 'nebula_ga_event_ajax' ) );
-
-            /*==========================
-                Nebula Visitor Data
-             ===========================*/
-
-            // TODO: Move into Nebula_Visitors or Nebula_Utilities_Visitors (I prefer Nebula_Visitors)
-
-            //Create Users Table with minimal default columns.
-            add_action('init', array( $this, 'create_visitors_table' ), 2); //Using init instead of admin_init so this triggers before check_nebula_id (below)
-
-            //Check if the Nebula ID exists on load and generate/store a new one if it does not.
-            add_action('init', array( $this, 'check_nebula_id' ), 11);
-
-            //Retrieve User Data
-            add_action('wp_ajax_nebula_ajax_get_visitor_data', array( $this, 'ajax_get_visitor_data' ) );
-            add_action('wp_ajax_nopriv_nebula_ajax_get_visitor_data', array( $this, 'ajax_get_visitor_data' ) );
-
-            //Vague Data - Only update if it doesn't already exist in the DB
-            add_action('wp_ajax_nebula_ajax_vague_visitor', array( $this, 'ajax_vague_visitor' ) );
-            add_action('wp_ajax_nopriv_nebula_ajax_vague_visitor', array( $this, 'ajax_vague_visitor' ) );
-
-            //Update Visitor Data
-            add_action('wp_ajax_nebula_ajax_update_visitor', array( $this, 'ajax_update_visitor' ) );
-            add_action('wp_ajax_nopriv_nebula_ajax_update_visitor', array( $this, 'ajax_update_visitor' ) );
-
-            //Append to Visitor Data
-            add_action('wp_ajax_nebula_ajax_append_visitor', array( $this, 'ajax_append_visitor' ) );
-            add_action('wp_ajax_nopriv_nebula_ajax_append_visitor', array( $this, 'ajax_append_visitor' ) );
-
-            //Increment Visitor Data
-            add_action('wp_ajax_nebula_ajax_increment_visitor', array( $this, 'ajax_increment_visitor' ) );
-            add_action('wp_ajax_nopriv_nebula_ajax_increment_visitor', array( $this, 'ajax_increment_visitor' ) );
-
-            /*==========================
-                Hubspot CRM Integration Functions
-
-                @TODO "Nebula" 0: Expand this functionality to include Salesforce and Marketo too.
-                https://github.com/chrisblakley/Nebula/issues/1182
-
-                Salesforce Documentation:
-                Marketo Documentation: http://developers.marketo.com/javascript-api/web-personalization/
-                    - JavaScript API, so this would be done right from main.js
-                    - So, nv() data would probably just make the call to the Marketo function too.
-             ===========================*/
-
-            // TODO: Move into Nebula_Hubspot_CRM or Nebula_Utilities_Hubspot_CRM (I prefer move this to an external plugin)
-
-            //Create/Update Contact in Hubspot CRM
-            add_action('wp_ajax_nebula_ajax_send_to_hubspot', array( $this, 'ajax_send_to_hubspot' ) );
-            add_action('wp_ajax_nopriv_nebula_ajax_send_to_hubspot', array( $this, 'ajax_send_to_hubspot' ) );
 
             //Fuzzy meta sub key finder (Used to query ACF nested repeater fields).
             //Example: 'key' => 'dates_%_start_date',
@@ -85,6 +65,7 @@ if( !class_exists( 'Nebula_Utilities' ) ) {
              ===========================*/
 
              // TODO: Move into Nebula_Sass or Nebula_Utilities_Sass (I prefer Nebula_Sass)
+             // TODO: Sass class needs a rewrite
 
             if ( nebula_option('scss', 'enabled') ){
 
@@ -94,16 +75,6 @@ if( !class_exists( 'Nebula_Utilities' ) ) {
 
                 add_action('init', array( $this, 'render_registered_scss' ), 99999); //Render registered plugins' Sass files (priority set to 99999 to run after all plugin registration)
             }
-
-            /*==========================
-             User Agent Parsing Functions/Helpers
-             ===========================*/
-
-             // TODO: Move into Nebula_Device_Detection or Nebula_Utilities_Device_Detection (I prefer Nebula_Device_Detection)
-
-            //Device Detection - https://github.com/piwik/device-detector
-            //Be careful when updating this library. DeviceDetector.php requires modification to work without Composer!
-            add_action('init', array( $this, 'device_detection' ) );
         }
 
         //Generate Session ID
@@ -359,921 +330,6 @@ if( !class_exists( 'Nebula_Utilities' ) ) {
 
             //concerned about large filesizes? Maybe only keep logs for the last 5000 lines or something?
             //append to log file.
-        }
-
-        //Create Users Table with minimal default columns.
-        public function nebula_create_visitors_table(){
-            if ( is_admin_page() && nebula_option('visitors_db') && isset($_GET['settings-updated']) && is_staff() ){ //Only trigger this in admin when Nebula Options are saved.
-                global $wpdb;
-
-                $visitors_table = $wpdb->query("SHOW TABLES LIKE 'nebula_visitors'");
-                if ( empty($visitors_table) ){
-                    $created = $wpdb->query("CREATE TABLE nebula_visitors (
-				id INT(11) NOT NULL AUTO_INCREMENT,
-				nebula_id TEXT NOT NULL,
-				known BOOLEAN NOT NULL DEFAULT FALSE,
-				email_address TEXT NOT NULL,
-				hubspot_vid INT(11) NOT NULL,
-				last_modified_date INT(12) NOT NULL DEFAULT 0,
-				last_session_id TEXT NOT NULL,
-				notable_poi TEXT NOT NULL,
-				score INT(5) NOT NULL DEFAULT 0,
-				score_mod INT(5) NOT NULL DEFAULT 0,
-				PRIMARY KEY (id)
-			) ENGINE = MyISAM;"); //Try InnoDB as engine? and add ROW_FORMAT=COMPRESSED ?
-                } else {
-                    nebula_remove_expired_visitors();
-                }
-            }
-        }
-
-        //Check if the Nebula ID exists on load and generate/store a new one if it does not.
-        public function check_nebula_id(){
-            if ( nebula_is_bot() || strpos(strtolower($_SERVER['HTTP_USER_AGENT']), 'wordpress') !== false ){
-                return false; //Don't add bots to the DB
-            }
-
-            $nebula_id = get_nebula_id();
-
-            if ( empty($nebula_id) ){ //If new user
-                generate_nebula_id();
-            }
-
-            $this->new_or_returning_visitor();
-        }
-
-        //Check if this visitor is new or returning using several factors
-        public function new_or_returning_visitor(){
-            if ( nebula_option('visitors_db') ){
-                $last_session_id = nebula_get_visitor_data('last_session_id'); //Check if this returning visitor exists in the DB (in case they were removed)
-                if ( empty($last_session_id) ){ //If the nebula_id is not in the DB already, treat it as a new user
-                    //Prevent duplicates for users blocking cookies or Google Analytics
-                    if ( strpos(ga_parse_cookie(), '-') !== false ){ //If GA CID was generated by Nebula
-                        global $wpdb;
-                        $unique_new_visitor = $wpdb->get_results("SELECT * FROM nebula_visitors WHERE ip_address = '" . sanitize_text_field($_SERVER['REMOTE_ADDR']) . "' AND user_agent = '" . sanitize_text_field($_SERVER['HTTP_USER_AGENT']) . "'");
-                        if ( !empty($unique_new_visitor) ){
-                            $unique_new_visitor = (array) $unique_new_visitor[0];
-                            if ( strpos($unique_new_visitor['ga_cid'], '-') !== false ){ //If that GA CID was also generated by Nebula
-                                generate_nebula_id($unique_new_visitor['nebula_id']); //Give this user the same Nebula ID
-                                //Update that row instead of inserting a new visitor.
-                                $this->update_visitor(array(
-                                    'first_session' => '0',
-                                    'notes' => 'This user tracked by IP and User Agent.',
-                                ));
-                            } else {
-                                $this->insert_visitor(array('first_session' => '0')); //The matching visitor row had a GA CID assigned (no dashes)
-                            }
-                        } else {
-                            $this->insert_visitor(array('first_session' => '0')); //No matching IP Address with same User Agent
-                        }
-                    } else {
-                        $this->insert_visitor(array('first_session' => '0'));
-                    }
-                } else {
-                    //Check if new session or another pageview in same session
-                    if ( session_id() != $last_session_id ){ //New session
-                        $update_data = array(
-                            'first_session' => '0',
-                            'current_session' => time(),
-                            'current_session_pageviews' => '1',
-                            'last_modified_date' => time(),
-                        );
-                        $this->increment_visitor(array('session_count'));
-                    } else { //Same session
-                        $update_data = array(
-                            'current_session' => time(),
-                            'last_modified_date' => time(),
-                        );
-                        $this->increment_visitor(array('current_session_pageviews'));
-                    }
-
-                    $this->update_visitor($update_data);
-                }
-            }
-        }
-
-        //Return the Nebula ID (or false)
-        public function get_nebula_id(){
-            if ( isset($_COOKIE['nid']) ){
-                return htmlentities(preg_replace('/[^a-zA-Z0-9\.]+/', '', $_COOKIE['nid']));
-            }
-
-            return false;
-        }
-
-        //Generate a unique Nebula ID and store it in a cookie and insert into DB
-        //Or force a specific Nebula ID
-        public function generate_nebula_id($force=null){
-            if ( !empty($force) ){
-                $_COOKIE['nid'] = $force;
-            } else {
-                $_COOKIE['nid'] = nebula_version('full') . '.' . bin2hex(openssl_random_pseudo_bytes(5)) . '.' . uniqid(); //Update to random_bytes() when moving to PHP7
-            }
-
-            $nid_expiration = strtotime('January 1, 2035'); //Note: Do not let this cookie expire past 2038 or it instantly expires.
-            setcookie('nid', $_COOKIE['nid'], $nid_expiration, COOKIEPATH, COOKIE_DOMAIN); //Store the Nebula ID as a cookie called "nid".
-
-            if ( nebula_option('visitors_db') ){
-                if ( empty($force) ){
-                    global $wpdb;
-                    $nebula_id_from_matching_ga_cid = $wpdb->get_results($wpdb->prepare("SELECT nebula_id FROM nebula_visitors WHERE ga_cid = '%s'", ga_parse_cookie())); //Check if the ga_cid exists, and if so use THAT nebula_id again
-                    if ( !empty($nebula_id_from_matching_ga_cid) ){
-                        $_COOKIE['nid'] = reset($nebula_id_from_matching_ga_cid[0]);
-                        setcookie('nid', $_COOKIE['nid'], $nid_expiration, COOKIEPATH, COOKIE_DOMAIN); //Update the Nebula ID cookie
-                    }
-                }
-            }
-
-            return $_COOKIE['nid'];
-        }
-
-        //Create necessary columns by comparing passed data to existing columns
-        public function nebula_visitors_create_missing_columns($all_data){
-            if ( nebula_option('visitors_db') ){
-                $existing_columns = nebula_visitors_existing_columns(); //Returns an array of current table column names
-
-                $needed_columns = array();
-                foreach ( $all_data as $column => $value ){
-                    if ( empty($value) ){
-                        $all_data[$column] = ''; //Convert null values to empty strings.
-                    }
-
-                    if ( !in_array($column, $existing_columns) ){ //If the column does not exist, add it to an array
-                        $needed_columns[] = $column;
-                    }
-                }
-
-                if ( !empty($needed_columns) ){
-                    global $wpdb;
-
-                    $alter_query = "ALTER TABLE nebula_visitors ";
-                    foreach ( $needed_columns as $column_name ){
-                        $column_name = sanitize_key($column_name);
-
-                        $sample_value = $all_data[$column_name];
-                        $data_type = 'TEXT'; //Default data type
-                        if ( is_int($sample_value) ){
-                            $data_type = 'INT(12)';
-                        } elseif ( strlen($sample_value) == 1 && ($sample_value === '1' || $sample_value === '0') ){
-                            $data_type = 'INT(7)';
-                        }
-
-                        $alter_query .= "ADD " . $column_name . " " . $data_type . " NOT NULL, "; //Prep each needed column into a query
-                    }
-
-                    $create_columns = $wpdb->query(rtrim($alter_query, ', ')); //Create the needed columns
-                }
-            }
-        }
-
-        //Retrieve User Data
-        public function ajax_get_visitor_data(){
-            if ( !wp_verify_nonce($_POST['nonce'], 'nebula_ajax_nonce') ){ die('Permission Denied.'); }
-            $column = sanitize_text_field($_POST['data']);
-            echo $this->get_visitor_data($column);
-            wp_die();
-        }
-
-        public function get_visitor_data($column){ //@TODO "Nebula" 0: Update to allow multiple datapoints to be accessed in one query.
-            if ( nebula_option('visitors_db') ){
-                $column = sanitize_key($column);
-
-                if ( $column == 'notes' ){
-                    return false;
-                }
-
-                $nebula_id = get_nebula_id();
-                if ( !empty($nebula_id) && !empty($column) ){
-                    global $wpdb;
-                    $requested_data = $wpdb->get_results($wpdb->prepare("SELECT " . $column . " FROM nebula_visitors WHERE nebula_id = '%s'", $nebula_id));
-
-                    if ( !empty($requested_data) && !empty($requested_data[0]) && strtolower(reset($requested_data[0])) != 'null' ){
-                        return reset($requested_data[0]); //@TODO "Nebula" 0: update so this could return multiple values
-                    }
-                }
-            }
-
-            return false;
-        }
-
-        //Vague Data - Only update if it doesn't already exist in the DB
-        public function ajax_low_visitor(){
-            if ( !wp_verify_nonce($_POST['nonce'], 'nebula_ajax_nonce') ){ die('Permission Denied.'); }
-            $data = $_POST['data'];
-            echo $this->vague_visitor($data);
-            wp_die();
-        }
-
-        public function vague_visitor($data=array()){
-            if ( nebula_option('visitors_db') ){
-                $data_to_send = array();
-                foreach ( $data as $column => $value ){
-                    $existing_value = $this->get_visitor_data($column);
-                    if ( empty($existing_value) ){ //If the requested data is empty/null, then update.
-                        $data_to_send[$column] = $value;
-                    }
-                }
-
-                if ( !empty($data_to_send) ){
-                    nebula_update_visitor($data_to_send);
-                }
-            }
-            return false;
-        }
-
-        //Update Visitor Data
-        public function ajax_update_visitor(){
-            if ( !wp_verify_nonce($_POST['nonce'], 'nebula_ajax_nonce') ){ die('Permission Denied.'); }
-            $data = $_POST['data'];
-            echo $this->update_visitor($data);
-            wp_die();
-        }
-
-        public function update_visitor($data=array(), $send_to_hubspot=true){ //$data is going to be array(column => value) or an array of arrays
-            if ( nebula_option('visitors_db') ){
-                global $wpdb;
-                $nebula_id = $this->get_nebula_id();
-
-                if ( !empty($nebula_id) ){
-                    $update_every_time = $this->visitor_data_update_everytime();
-                    $all_data = array_merge($update_every_time, $data); //Add any passed data
-
-                    //Check if the data should be sent to Hubspot
-                    if ( !empty($send_to_hubspot) ){
-                        $need_to_resend = false;
-                        $non_priority_columns = array('id', 'known', 'create_date', 'last_modified_date', 'initial_session', 'first_session', 'last_session_id', 'previous_session', 'current_session', 'current_session_pageviews', 'session_count', 'hubspot_vid', 'bot', 'expiration', 'score', 'ga_block', 'js_block', 'ad_block', 'nebula_session_id'); //These columns should not facilitate a Hubspot send by themselves.
-                        foreach ( $all_data as $column => $value ){
-                            if ( !in_array($column, $non_priority_columns) ){
-                                $need_to_resend = true;
-                            }
-                        }
-
-                        if ( empty($need_to_resend) ){
-                            $send_to_hubspot = false;
-                        }
-                    }
-
-                    nebula_visitors_create_missing_columns($all_data);
-                    $all_strings = array_fill(0, count($all_data), '%s'); //Create sanitization array
-
-                    //Update the visitor row
-                    $updated_visitor = $wpdb->update(
-                        'nebula_visitors',
-                        $all_data,
-                        array('nebula_id' => $nebula_id),
-                        $all_strings,
-                        array('%s')
-                    );
-
-                    if ( $updated_visitor === false ){ //If visitor does not exist in the table, create it with defaults and current data... might need to be true if its int(0) too, so maybe go back to empty($updated_visitor)
-                        $this->insert_visitor($all_data, $send_to_hubspot);
-                    } else {
-                        check_if_known($send_to_hubspot);
-                    }
-
-                    return true;
-                }
-            }
-
-            return false;
-        }
-
-        //Append to Visitor Data
-        public function ajax_append_visitor(){
-            if ( !wp_verify_nonce($_POST['nonce'], 'nebula_ajax_nonce') ){ die('Permission Denied.'); }
-            $data = $_POST['data']; //json_decode(stripslashes()); but its already an array... why?
-            echo $this->append_visitor($data);
-            wp_die();
-        }
-
-        public function append_visitor($data=array(), $send_to_hubspot=true){ //$data is going to be array(column => value) or an array of arrays
-            if ( nebula_option('visitors_db') ){
-                global $wpdb;
-                $nebula_id = get_nebula_id();
-
-                if ( !empty($nebula_id) && !empty($data) ){
-                    $this->update_visitor(array('last_modified_date' => time()));
-                    $this->visitors_create_missing_columns($data);
-
-                    $append_query = "UPDATE nebula_visitors ";
-                    foreach ( $data as $column => $value ){
-                        $column = sanitize_key($column);
-
-                        $value = sanitize_text_field($value);
-                        $append_query .= "SET " . $column . " = CONCAT_WS(',', NULLIF(" . $column . ", ''), '" . $value . "'),"; //how to further prepare/sanitize this value? not sure how many %s are needed...
-                    }
-                    $append_query = rtrim($append_query, ', ');
-                    $append_query .= "WHERE nebula_id = '" . $nebula_id . "'";
-
-                    if ( strpos(str_replace(' ', '', $append_query), 'nebula_visitorsWHERE') === false ){
-                        $appended_visitor = $wpdb->query($append_query); //currently working on this
-
-                        if ( $appended_visitor === false ){ //If visitor does not exist in the table, create it with defaults and current data. might need to be true if its int(0) too, so maybe go back to empty($updated_visitor)
-                            $this->insert_visitor($data, $send_to_hubspot);
-                        } else {
-                            check_if_known($send_to_hubspot);
-                        }
-
-                        return true;
-                    }
-                }
-            }
-
-            return false;
-        }
-
-        //Increment Visitor Data
-        public function ajax_increment_visitor(){
-            if ( !wp_verify_nonce($_POST['nonce'], 'nebula_ajax_nonce') ){ die('Permission Denied.'); }
-            $data = $_POST['data'];
-            echo nebula_increment_visitor($data);
-            wp_die();
-        }
-
-        public function increment_visitor($data){ //Data should be an array of columns to increment
-            if ( nebula_option('visitors_db') ){
-                global $wpdb;
-                $nebula_id = get_nebula_id();
-
-                if ( is_string($data) ){
-                    $data = array($data);
-                }
-
-                //@TODO "Nebula" 0: echo here to see if this gets triggered multiple times (when it should only get triggered once)
-
-                if ( !empty($nebula_id) && !empty($data) ){
-                    $increment_query = "UPDATE nebula_visitors ";
-                    foreach ( $data as $column ){
-                        $column = sanitize_key($column);
-                        $increment_query .= "SET " . $column . " = " . $column . "+1, ";
-                    }
-                    $increment_query = rtrim($increment_query, ', ') . ' ';
-                    $increment_query .= "WHERE nebula_id = '" . $nebula_id . "'";
-
-                    if ( strpos(str_replace(' ', '', $increment_query), 'nebula_visitorsWHERE') === false ){
-                        $incremented_visitor = $wpdb->query($increment_query);
-                        if ( $incremented_visitor === false ){ //If visitor does not exist in the table, create it with defaults and current data... might need to be true if its int(0) too, so maybe go back to empty($updated_visitor)
-                            return false;
-                        }
-
-                        return true;
-                    }
-                }
-            }
-
-            return false;
-        }
-
-        //Insert visitor into table with all default detections
-        public function insert_visitor($data=array(), $send_to_hubspot=true){
-            if ( nebula_option('visitors_db') ){
-                $nebula_id = get_nebula_id();
-                if ( !empty($nebula_id) ){
-                    global $wpdb;
-
-                    $defaults = array(
-                        'nebula_id' => $nebula_id,
-                        'ga_cid' => ga_parse_cookie(), //Will be UUID on first visit then followed up with actual GA CID via AJAX (if available)
-                        'known' => '0',
-                        'ip_address' => sanitize_text_field($_SERVER['REMOTE_ADDR']),
-                        'create_date' => time(),
-                        'hubspot_vid' => false,
-                        'score' => '0',
-                        'notes' => '',
-                        'referer' => ( isset($_SERVER['HTTP_REFERER']) )? sanitize_text_field($_SERVER['HTTP_REFERER']) : '',
-                        'first_session' => '1',
-                        'session_count' => '1',
-                        'prev_session' => '0',
-                        'last_session_id' => session_id(),
-                        'nebula_session_id' => nebula_session_id(),
-                        'current_session' => time(),
-                        'current_session_pageviews' => '1',
-                        'prerendered' => '0',
-                        'page_visibility_hidden' => '0',
-                        'page_visibility_visible' => '0',
-                        'ga_block' => '0',
-                        'page_not_found' => '0',
-                        'no_search_results' => '0',
-                        'external_links' => '0',
-                        'non_linked_click' => '0',
-                        'copied_text' => '0',
-                        'html_errors' => '0',
-                        'js_errors' => '0',
-                        'css_errors' => '0',
-                        'ajax_errors' => '0',
-                        'page_suggestion_clicks' => '0',
-                        'infinite_query_loads' => '0',
-                        'score_mod' => '0',
-                    );
-
-                    //Attempt to detect IP Geolocation data using https://freegeoip.net/
-                    if ( nebula_option('ip_geolocation') && nebula_is_available('http://freegeoip.net') ){
-                        $response = wp_remote_get('http://freegeoip.net/json/' . $_SERVER['REMOTE_ADDR']);
-                        if ( !is_wp_error($response) ){
-                            $ip_geo_data = json_decode($response['body']);
-                            if ( !empty($ip_geo_data) ){
-                                $defaults['ip_country'] = sanitize_text_field($ip_geo_data->country_name);
-                                $defaults['ip_region'] = sanitize_text_field($ip_geo_data->region_name);
-                                $defaults['ip_city'] = sanitize_text_field($ip_geo_data->city);
-                                $defaults['ip_zip'] = sanitize_text_field($ip_geo_data->zip_code);
-                            }
-                        } else {
-                            set_transient('nebula_site_available_' . str_replace('.', '_', nebula_url_components('hostname', 'http://freegeoip.net/json/')), 'Unavailable', 60*5); //5 minute expiration
-                        }
-                    }
-
-                    $defaults = nebula_visitor_data_update_everytime($defaults);
-
-                    $all_data = array_merge($defaults, $data); //Add any passed data
-                    nebula_visitors_create_missing_columns($all_data);
-                    $all_strings = array_fill(0, count($all_data), '%s'); //Create sanitization array
-
-                    $wpdb->insert('nebula_visitors', $all_data, $all_strings); //Insert a row with all the default (and passed) sanitized values.
-
-                    check_if_known($send_to_hubspot);
-
-                    nebula_remove_expired_visitors();
-                    return true;
-                }
-            }
-
-            return false;
-        }
-
-        //Update certain visitor data everytime.
-        //Pass the existing data array to this to append to it (otherwise a new array is created)!
-        public function visitor_data_update_everytime($defaults=array()){
-            $defaults['ga_cid'] = ga_parse_cookie();
-            $defaults['notable_poi'] = nebula_poi();
-            $defaults['last_modified_date'] = time();
-            $defaults['nebula_session_id'] = nebula_session_id();
-            $defaults['score'] = nebula_calculate_visitor_score(); //Try to limit this (without doing a DB query to check)
-
-            //Avoid ternary operators to prevent overwriting existing data (like manual DB entries)
-
-            //Check for nv_ query parameters
-            $query_strings = parse_str($_SERVER['QUERY_STRING']);
-            if ( !empty($query_strings) ){
-                foreach ( $query_strings as $key => $value ){
-                    if ( strpos($key, 'nv_') === 0 ){
-                        if ( empty($value) ){
-                            $value = 'true';
-                        }
-                        $defaults[sanitize_key(substr($key, 3))] = sanitize_text_field(str_replace('+', ' ', urldecode($value)));
-                    }
-                }
-            }
-
-            //Logged-in User Data
-            if ( is_user_logged_in() ){
-                $defaults['wp_user_id'] = get_current_user_id();
-
-                $user = get_userdata(get_current_user_id());
-                if ( !empty($user) ){
-                    //Default WordPress user info
-                    if ( !empty($user->roles) ){
-                        $defaults['wp_role'] = sanitize_text_field($user->roles[0]);
-                    }
-                    if ( !empty($user->user_firstname) ){
-                        $defaults['first_name'] = sanitize_text_field($user->user_firstname);
-                    }
-                    if ( !empty($user->user_lastname) ){
-                        $defaults['last_name'] = sanitize_text_field($user->user_lastname);
-                    }
-                    if ( !empty($user->user_firstname) && !empty($user->user_lastname) ){
-                        $defaults['full_name'] = sanitize_text_field($user->user_firstname . ' ' . $user->user_lastname);
-                    }
-                    if ( !empty($user->user_email) ){
-                        $defaults['email_address'] = sanitize_text_field($user->user_email);
-                    }
-                    if ( !empty($user->user_login) ){
-                        $defaults['username'] = sanitize_text_field($user->user_login);
-                    }
-
-                    //Custom user fields
-                    if ( get_user_meta($user->ID, 'headshot_url', true) ){
-                        $defaults['photo'] = sanitize_text_field(get_user_meta($user->ID, 'headshot_url', true));
-                    }
-                    if ( get_user_meta($user->ID, 'jobtitle', true) ){
-                        $defaults['job_title'] = sanitize_text_field(get_user_meta($user->ID, 'jobtitle', true));
-                    }
-                    if ( get_user_meta($user->ID, 'jobcompany', true) ){
-                        $defaults['company'] = sanitize_text_field(get_user_meta($user->ID, 'jobcompany', true));
-                    }
-                    if ( get_user_meta($user->ID, 'jobcompanywebsite', true) ){
-                        $defaults['company_website'] = sanitize_text_field(get_user_meta($user->ID, 'jobcompanywebsite', true));
-                    }
-                    if ( get_user_meta($user->ID, 'phonenumber', true) ){
-                        $defaults['phone_number'] = sanitize_text_field(get_user_meta($user->ID, 'phonenumber', true));
-                    }
-                    if ( get_user_meta($user->ID, 'usercity', true) ){
-                        $defaults['city'] = sanitize_text_field(get_user_meta($user->ID, 'usercity', true));
-                    }
-                    if ( get_user_meta($user->ID, 'userstate', true) ){
-                        $defaults['state_name'] = sanitize_text_field(get_user_meta($user->ID, 'userstate', true));
-                    }
-                }
-            }
-
-            //Campaign Data
-            if ( isset($_GET['utm_campaign']) ){
-                $defaults['utm_campaign'] = sanitize_text_field($_GET['utm_campaign']);
-            }
-            if ( isset($_GET['utm_medium']) ){
-                $defaults['utm_medium'] = sanitize_text_field($_GET['utm_medium']);
-            }
-            if ( isset($_GET['utm_source']) ){
-                $defaults['utm_source'] = sanitize_text_field($_GET['utm_source']);
-            }
-            if ( isset($_GET['utm_content']) ){
-                $defaults['utm_content'] = sanitize_text_field($_GET['utm_content']);
-            }
-            if ( isset($_GET['utm_term']) ){
-                $defaults['utm_term'] = sanitize_text_field($_GET['utm_term']);
-            }
-
-            //Device information
-            $defaults['user_agent'] = sanitize_text_field($_SERVER['HTTP_USER_AGENT']);
-            $defaults['device_form_factor'] = nebula_get_device('formfactor');
-            $defaults['device_full'] = nebula_get_device('full');
-            $defaults['device_brand'] = nebula_get_device('brand');
-            $defaults['device_model'] = nebula_get_device('model');
-            $defaults['device_type'] = nebula_get_device('type');
-            $defaults['os_full'] = nebula_get_os('full');
-            $defaults['os_name'] = nebula_get_os('name');
-            $defaults['os_version'] = nebula_get_os('version');
-            $defaults['browser_full'] = nebula_get_browser('full');
-            $defaults['browser_name'] = nebula_get_browser('name');
-            $defaults['browser_version'] = nebula_get_browser('version');
-            $defaults['browser_engine'] = nebula_get_browser('engine');
-            $defaults['browser_type'] = nebula_get_browser('type');
-
-            return $defaults;
-        }
-
-        //Calculate and update the visitor with a score
-        public function calculate_visitor_score($id=null){
-            if ( nebula_option('visitors_db') ){
-                global $wpdb;
-
-                //Which visitor data to calculate for?
-                if ( !empty($id) && current_user_can('manage_options') ){ //If an ID is passed and the user is an admin, use the passed visitor ID (Note: ID is from the visitor DB- not WordPress ID).
-                    $this_visitor = $wpdb->get_results("SELECT * FROM nebula_visitors WHERE id = '" . $id . "'");
-                } else { //Else, use the current visitor
-                    $nebula_id = get_nebula_id();
-                    if ( !empty($nebula_id) ){
-                        $this_visitor = $wpdb->get_results("SELECT * FROM nebula_visitors WHERE nebula_id = '" . $nebula_id . "'");
-                    }
-                }
-
-                if ( $this_visitor ){
-                    $this_visitor = (array) $this_visitor[0];
-
-                    //A score of 100+ will prevent deletion
-                    $point_values = array(
-                        'notable_poi' => 100,
-                        'known' => 100,
-                        'hubspot_vid' => 100,
-                        'notes' => 25,
-                        'notable_download' => 10,
-                        'pdf_view' => 10,
-                        'internal_search' => 20,
-                        'contact_method' => 75,
-                        'ecommerce_addtocart' => 75,
-                        'ecommerce_checkout' => 75,
-                        'engaged_reader' => 10,
-                        'contact_funnel' => 25,
-                        'street_full' => 75,
-                        'geo_latitude' => 75,
-                        'video_play' => 10,
-                        'video_engaged' => 25,
-                        'video_finished' => 25,
-                        'fb_like' => 20,
-                        'fb_share' => 20,
-                        'score_mod' => $this_visitor['score_mod'],
-                    );
-
-                    $score = ( $this_visitor['session_count'] > 1 )? $this_visitor['session_count'] : 0; //Start the score at the session count if higher than 1
-                    foreach ( $this_visitor as $column => $value ){
-                        if ( array_key_exists($column, $point_values) && !empty($value) ){
-                            if ( $column == 'notes' && $value == 'This user tracked by IP and User Agent.' ){
-                                continue;
-                            }
-                            $score += $point_values[$column];
-                        }
-                    }
-
-                    return $score;
-                }
-            }
-
-            return false;
-        }
-
-        //Remove expired visitors from the DB
-        //This is only ran when Nebula Options are saved, and when *new* visitors are inserted.
-        public function remove_expired_visitors(){
-            if ( nebula_option('visitors_db') ){
-                global $wpdb;
-                $expiration_length = time()-2592000; //30 days
-                $wpdb->query($wpdb->prepare("DELETE FROM nebula_visitors WHERE last_modified_date < %d AND known = %d AND score < %d", $expiration_length, 0, 100));
-            }
-        }
-
-        //Look up what columns currently exist in the nebula_visitors table.
-        public function visitors_existing_columns(){
-            global $wpdb;
-            return $wpdb->get_col("SHOW COLUMNS FROM nebula_visitors", 0); //Returns an array of current table column names
-        }
-
-        //Query email address or Hubspot VID to see if the user is known
-        public function check_if_known($send_to_hubspot=true){
-            if ( nebula_option('visitors_db') ){
-                global $wpdb;
-                $nebula_id = get_nebula_id();
-
-                $known_visitor = $wpdb->get_results("SELECT * FROM nebula_visitors WHERE nebula_id LIKE '" . $nebula_id . "' AND (email_address REGEXP '.+@.+\..+' OR hubspot_vid REGEXP '^\d+$')");
-                if ( !empty($known_visitor) ){
-                    if ( !is_known() ){
-                        nebula_update_visitor(array('known' => '1')); //Update to known visitor (if previously unknown)
-                    }
-
-                    $known_visitor_data = (array) $known_visitor[0];
-                    if ( !empty($send_to_hubspot) ){
-                        nebula_prep_data_for_hubspot_crm_delivery($known_visitor_data);
-                    }
-                    return true;
-                }
-            }
-
-            return false;
-        }
-
-        //Lookup if this visitor is known
-        public function is_known(){
-            $known = nebula_get_visitor_data('known');
-            if ( !empty($known) && ($known === 1 || $known === '1') ){ //@TODO "Nebula" 0: Figure out which of these is best
-                return true;
-            }
-
-            return false;
-        }
-
-        //Prepare Nebula Visitor data to be sent to Hubspot CRM
-        //This includes skipping empty fields, ignoring certain fields, and renaming others.
-        public function prep_data_for_hubspot_crm_delivery($data){
-            if ( nebula_option('hubspot_api') ){
-                $data_for_hubspot = array();
-                if ( !empty($data['hubspot_vid']) ){
-                    $data_for_hubspot['hubspot_vid'] = $data['hubspot_vid'];
-                }
-                if ( !empty($data['email_address']) ){
-                    $data_for_hubspot['email_address'] = $data['email_address'];
-                }
-                $data_for_hubspot['properties'] = array();
-
-                $ignore_columns = array('id', 'known', 'last_modified_date', 'first_session', 'prev_session', 'last_session_id', 'current_session', 'hubspot_vid', 'bot', 'current_session_pageviews', 'score', 'expiration', 'street_number', 'street_name', 'zip_suffix', 'zip_full');
-                $rename_columns = array(
-                    'ip_address' => 'nebula_ip',
-                    'ip_city' => 'nebula_ip_city',
-                    'ip_region' => 'nebula_ip_region',
-                    'ip_country' => 'nebula_ip_country',
-                    'street_full' => 'address',
-                    'state_abbr' => 'state',
-                    'country_name' => 'country',
-                    'zip_code' => 'zip',
-                    'first_name' => 'firstname',
-                    'last_name' => 'lastname',
-                    'email_address' => 'email',
-                    'phone_number' => 'phone',
-                );
-
-                foreach ( $data as $column => $value ){
-                    //Skip empty column values
-                    if ( empty($value) ){
-                        continue;
-                    }
-
-                    //Ignore unnecessary columns
-                    if ( in_array($column, $ignore_columns) ){
-                        continue;
-                    }
-
-                    //Rename certain columns to Hubspot CRM notation
-                    if ( array_key_exists($column, $rename_columns) ){
-                        $column = $rename_columns[$column];
-                    }
-
-                    //Add the column/value to the Hubspot data array
-                    $data_for_hubspot['properties'][] = array(
-                        'property' => $column,
-                        'value' => $value
-                    );
-                }
-
-                nebula_send_to_hubspot($data_for_hubspot);
-            }
-        }
-
-        // Hubspot CRM
-
-        //Send data to Hubspot CRM via PHP curl
-        public function hubspot_curl($url, $content=null){
-            $sep = ( strpos($url, '?') === false )? '?' : '&';
-            $get_url = $url . $sep . 'hapikey=' . nebula_option('hubspot_api');
-
-            if ( !empty($content) ){
-                /*
-                    @TODO "Nebula" 0: 409 Conflict response happening. Was probably happening with cURL and just never noticed.
-                        - Because the fields already exist, Hubspot is responding with "409 Conflict".
-                        - This happens ~14 times since each property is sent individually.
-                        - I'm pretty sure the data is still transferring just fine.
-                        - Query Monitor is going red due to the 400-level response.
-                        - This is a Hubspot CRM issue, not WordPress or Nebula (as far as I can tell)
-                */
-
-                $response = wp_remote_post($get_url, array(
-                    'headers'  => array('Content-Type' => 'application/json'),
-                    'body' => $content,
-                ));
-            } else {
-                $response = wp_remote_get($get_url);
-            }
-
-            if ( !is_wp_error($response) ){
-                return $response['body'];
-            }
-
-            set_transient('nebula_site_available_' . str_replace('.', '_', nebula_url_components('hostname', $get_url)), 'Unavailable', 60*5); //5 minute expiration
-            return false;
-        }
-
-        //Get all existing Hubspot CRM contact properties in the Nebula group
-        public function get_hubspot_properties(){
-            $all_hubspot_properties = nebula_hubspot_curl('https://api.hubapi.com/contacts/v2/properties');
-            $all_hubspot_properties = json_decode($all_hubspot_properties, true);
-
-            $existing_nebula_properties = array();
-            foreach ( $all_hubspot_properties as $property ){
-                if ( $property['groupName'] == 'nebula' ){
-                    $existing_nebula_properties[] = $property['name'];
-                }
-            }
-
-            return $existing_nebula_properties;
-        }
-
-        //Create Custom Properties
-        public function create_hubspot_properties($columns=null){
-            if ( nebula_option('hubspot_portal') ){
-                //Create the Nebula group of properties
-                $content = '{
-	        "name": "nebula",
-	        "displayName": "Nebula",
-	        "displayOrder": 5
-	    }';
-                nebula_hubspot_curl('http://api.hubapi.com/contacts/v2/groups?portalId=' . nebula_option('hubspot_portal'), $content);
-
-                //Get an array of all existing Hubspot CRM contact properties
-                $existing_nebula_properties = get_nebula_hubspot_properties();
-
-                //Create Nebula IP custom property within the Nebula group
-                if ( !in_array('nebula_ip', $existing_nebula_properties) ){
-                    $content = '{
-				"name": "nebula_ip",
-				"label": "IP Address (Nebula)",
-				"description": "The IP address.",
-				"groupName": "nebula",
-				"type": "string",
-				"fieldType": "text",
-				"formField": true,
-				"displayOrder": 6,
-				"options": []
-			}';
-                    nebula_hubspot_curl('https://api.hubapi.com/contacts/v2/properties', $content);
-                }
-
-                //Create a property from the passed array
-                //@todo "Nebula" 0: This loop is accounting for 1 second of server time (about 0.14s per call)! Need to find a way to optimize...
-                if ( !empty($columns) ){
-                    if ( is_string($columns) ){
-                        $columns = array($columns);
-                    }
-                    foreach ( $columns as $column ){
-                        $column_label = ucwords(str_replace('_', ' ', $column));
-                        if ( !in_array($column, $existing_nebula_properties) ){
-                            //Create custom property within the Nebula group
-                            $content = '{
-						"name": "' . $column . '",
-						"label": "' . $column_label . '",
-						"description": "' . $column_label . ' (Parsed from Visitor DB).",
-						"groupName": "nebula",
-						"type": "string",
-						"fieldType": "text",
-						"formField": true,
-						"displayOrder": 6,
-						"options": []
-					}';
-                            nebula_hubspot_curl('https://api.hubapi.com/contacts/v2/properties', $content);
-                        }
-                    }
-                }
-            }
-        }
-
-        //Create/Update Contact in Hubspot CRM
-        public function ajax_send_to_hubspot(){
-            if ( !wp_verify_nonce($_POST['nonce'], 'nebula_ajax_nonce') ){ die('Permission Denied.'); }
-
-            $data = array(
-                'properties' => $_POST['properties'],
-            );
-
-            echo nebula_send_to_hubspot($data);
-            wp_die();
-        }
-
-        public function send_to_hubspot($data=array()){
-            if ( nebula_option('hubspot_api') ){
-                //Determine if we'll be using email_address or hubspot_vid for our send. Prefer VID.
-                if ( empty($data['hubspot_vid']) && empty($data['email_address']) ){ //If calling from AJAX we must lookup VID or Email Address
-                    global $wpdb;
-                    $nebula_id = get_nebula_id();
-                    $vid_and_email = $wpdb->get_results("SELECT hubspot_vid, email_address FROM nebula_visitors WHERE nebula_id LIKE '" . $nebula_id . "' AND email_address <> '' OR hubspot_vid <> ''"); //here
-                    $vid_and_email = (array) $vid_and_email[0];
-                    $hubspot_vid = $vid_and_email['hubspot_vid'];
-                    $email_address = $vid_and_email['email_address'];
-                } else { //Calling directly from another PHP function
-                    if ( !empty($data['hubspot_vid']) ){
-                        $hubspot_vid = $data['hubspot_vid'];
-                    }
-                    if ( !empty($data['email_address']) ){
-                        $email_address = $data['email_address'];
-                    }
-                }
-
-                if ( !empty($hubspot_vid) || !empty($email_address) ){ //If visitor has hubspot_vid or email_address
-                    //Create the properties array
-                    $content = array('properties' => array());
-                    $needed_properties = array();
-
-                    //Loop through provided properties
-                    foreach ( $data['properties'] as $group ){
-                        $needed_properties[] = $group['property'];
-
-                        $content['properties'][] = array(
-                            'property' => $group['property'],
-                            'value' => $group['value']
-                        );
-                    }
-
-                    nebula_create_hubspot_properties($needed_properties); //Check and create existing properties
-
-                    if ( !empty($hubspot_vid) ){
-                        $response = nebula_hubspot_curl('https://api.hubapi.com/contacts/v1/contact/vid/' . $hubspot_vid . '/profile', json_encode($content)); //Update the existing contact using their VID
-
-                        if ( strpos($response, 'error') != false ){ //There was an error
-                            return false;
-                        } elseif ( is_string($response) && $response == '' ){ //This API reponse is simply a 200 when it succeeds (empty string).
-                            nebula_update_visitor(array('hubspot_vid' => $hubspot_vid), false); //Update visitor withour re-sending to Hubspot CRM
-                            return $hubspot_vid;
-                        }
-                    } else {
-                        $response = nebula_hubspot_curl('https://api.hubapi.com/contacts/v1/contact/createOrUpdate/email/' . $email_address . '/', json_encode($content)); //Create or update the contact using their Email
-                        if ( strpos($response, 'error') != false ){ //There was an error
-                            return false;
-                        } else {
-                            $response = (array) json_decode($response);
-                            nebula_update_visitor(array('hubspot_vid' => $response['vid']), false); //Update visitor withour re-sending to Hubspot CRM
-                            return $response['vid'];
-                        }
-                    }
-                }
-            }
-
-            return false;
-        }
-
-        //Get contact data from Hubspot
-        //Set this to a variable to avoid multiple calls, then parse it like this: $hubspot_data['properties']['firstname']['value']
-        public function get_hubspot_contact($vid=null, $property=''){
-            if ( empty($vid) ){
-                $vid = $this->get_visitor_data('hubspot_vid');
-                if ( empty($vid) ){
-                    return false;
-                }
-            }
-
-            if ( !empty($property) ){
-                $property = '&property=' . $property;
-            }
-
-            $response = wp_remote_get('https://api.hubapi.com/contacts/v1/contact/vid/' . $vid . '/profile?hapikey=' . nebula_option('hubspot_api') . $property);
-            if ( is_wp_error($response) ){
-                set_transient('nebula_site_available_' . str_replace('.', '_', nebula_url_components('hostname', 'https://api.hubapi.com/')), 'Unavailable', 60*5); //5 minute expiration
-                return false;
-            }
-            return json_decode($response['body'], true);
         }
 
         //Detect Notable POI
@@ -2323,380 +1379,6 @@ if( !class_exists( 'Nebula_Utilities' ) ) {
 
             preg_match('/\$' . $color_search . ': (\S*)(;| !default;)/', $scss_variables, $matches);
             return $matches[1];
-        }
-
-        /*==========================
-         User Agent Parsing Functions/Helpers
-         ===========================*/
-
-        // TODO: If we creates the class Nebula_Device_Detection, this method should be detect()
-        public function device_detection(){
-            if ( nebula_option('device_detection') ){
-                require_once(get_template_directory() . '/includes/libs/device-detector/DeviceDetector.php'); // TODO: If we move classes to includes, includes/libs should go on libs
-                $GLOBALS["device_detect"] = new DeviceDetector($_SERVER['HTTP_USER_AGENT']);
-                $GLOBALS["device_detect"]->discardBotInformation(); //If called, getBot() will only return true if a bot was detected (speeds up detection a bit)
-                $GLOBALS["device_detect"]->parse();
-            }
-        }
-
-        //Boolean return if the user's device is mobile.
-        public function is_mobile(){
-            $override = apply_filters('pre_nebula_is_mobile', false);
-            if ( $override !== false ){return $override;}
-
-            if ( nebula_option('device_detection') ){
-                if ( $GLOBALS["device_detect"]->isMobile() ){
-                    return true;
-                }
-            }
-
-            global $is_iphone;
-            if ( $is_iphone ){
-                return true;
-            }
-
-            return false;
-        }
-
-        //Boolean return if the user's device is a tablet.
-        public function is_tablet(){
-            $override = apply_filters('pre_nebula_is_tablet', false);
-            if ( $override !== false ){return $override;}
-
-            if ( nebula_option('device_detection') ){
-                if ( $GLOBALS["device_detect"]->isTablet() ){
-                    return true;
-                }
-            }
-
-            return false;
-        }
-
-        //Boolean return if the user's device is a desktop.
-        public function is_desktop(){
-            $override = apply_filters('pre_nebula_is_desktop', false);
-            if ( $override !== false ){return $override;}
-
-            if ( nebula_option('device_detection') ){
-                if ( $GLOBALS["device_detect"]->isDesktop() ){
-                    return true;
-                }
-            }
-
-            return false;
-        }
-
-        //Returns the requested information of the operating system of the user's device.
-        public function get_os($info='full'){
-            $override = apply_filters('pre_nebula_get_os', false, $info);
-            if ( $override !== false ){return $override;}
-
-            if ( nebula_option('device_detection') ){
-                $os = $GLOBALS["device_detect"]->getOs();
-                switch ( strtolower($info) ){
-                    case 'full':
-                        return $os['name'] . ' ' . $os['version'];
-                        break;
-                    case 'name':
-                        return $os['name'];
-                        break;
-                    case 'version':
-                        return $os['version'];
-                        break;
-                    default:
-                        return false;
-                        break;
-                }
-            }
-
-            global $is_iphone;
-            switch ( strtolower($info) ){
-                case 'full':
-                case 'name':
-                    if ( $is_iphone ){
-                        return 'ios';
-                    }
-                    break;
-                default:
-                    return false;
-                    break;
-            }
-        }
-
-        //Check to see how the operating system version of the user's device compares to a passed version number.
-        public function is_os($os=null, $version=null, $comparison='=='){
-            $override = apply_filters('pre_nebula_is_os', false, $os, $version, $comparison);
-            if ( $override !== false ){return $override;}
-
-            if ( nebula_option('device_detection') ){
-                if ( empty($os) ){
-                    trigger_error('nebula_is_os requires a parameter of requested operating system.');
-                    return false;
-                }
-
-                switch ( strtolower($os) ){
-                    case 'macintosh':
-                        $os = 'mac';
-                        break;
-                    case 'win':
-                        $os = 'windows';
-                        break;
-                }
-
-                $actual_os = $GLOBALS["device_detect"]->getOs();
-                $actual_version = explode('.', $actual_os['version']);
-                $version_parts = explode('.', $version);
-                if ( strpos(strtolower($actual_os['name']), strtolower($os)) !== false ){
-                    if ( !empty($version) ){
-                        if ( nebula_compare_operator($actual_version[0], $version_parts[0], $comparison) ){ //If major version matches
-                            if ( $version_parts[1] && $version_parts[1] != 0 ){ //If minor version exists and is not 0
-                                if ( nebula_compare_operator($actual_version[1], $version_parts[1], $comparison) ){ //If minor version matches
-                                    return true;
-                                } else {
-                                    return false;
-                                }
-                            } else {
-                                return true;
-                            }
-                        }
-                    } else {
-                        return true;
-                    }
-                }
-            }
-
-            return false;
-        }
-
-        //Returns the requested information of the model of the user's device.
-        public function get_device($info='model'){
-            $override = apply_filters('pre_nebula_get_device', false, $info);
-            if ( $override !== false ){return $override;}
-
-            if ( nebula_option('device_detection') ){
-                $info = str_replace(' ', '', $info);
-                switch ( strtolower($info) ){
-                    case 'full':
-                        return $GLOBALS["device_detect"]->getBrandName() . ' ' . $GLOBALS["device_detect"]->getModel();
-                        break;
-                    case 'brand':
-                    case 'brandname':
-                    case 'make':
-                        return $GLOBALS["device_detect"]->getBrandName();
-                        break;
-                    case 'model':
-                    case 'version':
-                    case 'name':
-                        return $GLOBALS["device_detect"]->getModel();
-                        break;
-                    case 'type':
-                        return $GLOBALS["device_detect"]->getDeviceName();
-                        break;
-                    case 'formfactor':
-                        if ( nebula_is_mobile() ){
-                            return 'mobile';
-                        } elseif ( nebula_is_tablet() ){
-                            return 'tablet';
-                        } else {
-                            return 'desktop';
-                        }
-                    default:
-                        return false;
-                        break;
-                }
-            }
-
-            global $is_iphone;
-            $info = str_replace(' ', '', $info);
-            switch ( strtolower($info) ){
-                case 'brand':
-                case 'brandname':
-                case 'make':
-                case 'model':
-                case 'name':
-                case 'type':
-                    if ( $is_iphone ){
-                        return 'iphone';
-                    }
-                    break;
-                case 'formfactor':
-                    if ( $is_iphone ){
-                        return 'mobile';
-                    }
-                    break;
-                default:
-                    return false;
-                    break;
-            }
-        }
-
-        //Returns the requested information of the browser being used.
-        public function get_client($info){ return get_browser($info); }
-        public function get_browser($info='name'){
-            $override = apply_filters('pre_nebula_get_browser', false, $info);
-            if ( $override !== false ){return $override;}
-
-            if ( nebula_option('device_detection') ){
-                $client = $GLOBALS["device_detect"]->getClient();
-                switch ( strtolower($info) ){
-                    case 'full':
-                        return $client['name'] . ' ' . $client['version'];
-                        break;
-                    case 'name':
-                    case 'browser':
-                    case 'client':
-                        return $client['name'];
-                        break;
-                    case 'version':
-                        return $client['version'];
-                        break;
-                    case 'engine':
-                        return $client['engine'];
-                        break;
-                    case 'type':
-                        return $client['type'];
-                        break;
-                    default:
-                        return false;
-                        break;
-                }
-            }
-
-            global $is_gecko, $is_IE, $is_opera, $is_safari, $is_chrome;
-            switch ( strtolower($info) ){
-                case 'full':
-                case 'name':
-                case 'browser':
-                case 'client':
-                    if ( $is_IE ){return 'internet explorer';}
-                    elseif ( $is_opera ){return 'opera';}
-                    elseif ( $is_safari ){return 'safari';}
-                    elseif ( $is_chrome ){return 'chrome';}
-                    break;
-                case 'engine':
-                    if ( $is_gecko ){return 'gecko';}
-                    elseif ( $is_safari ){return 'webkit';}
-                    elseif ( $is_IE ){return 'trident';}
-                    break;
-                default:
-                    return false;
-                    break;
-            }
-        }
-
-        //Check to see how the browser version compares to a passed version number.
-        public function is_browser($browser=null, $version=null, $comparison='=='){
-            $override = apply_filters('pre_nebula_is_browser', false, $browser, $version, $comparison);
-            if ( $override !== false ){return $override;}
-
-            if ( nebula_option('device_detection') ){
-                if ( empty($browser) ){
-                    trigger_error('nebula_is_browser requires a parameter of requested browser.');
-                    return false;
-                }
-
-                switch ( strtolower($browser) ){
-                    case 'ie':
-                        $browser = 'internet explorer';
-                        break;
-                    case 'ie7':
-                        $browser = 'internet explorer';
-                        $version = '7';
-                        break;
-                    case 'ie8':
-                        $browser = 'internet explorer';
-                        $version = '8';
-                        break;
-                    case 'ie9':
-                        $browser = 'internet explorer';
-                        $version = '9';
-                        break;
-                    case 'ie10':
-                        $browser = 'internet explorer';
-                        $version = '10';
-                        break;
-                    case 'ie11':
-                        $browser = 'internet explorer';
-                        $version = '11';
-                        break;
-                }
-
-                $actual_browser = $GLOBALS["device_detect"]->getClient();
-                $actual_version = explode('.', $actual_browser['version']);
-                $version_parts = explode('.', $version);
-                if ( strpos(strtolower($actual_browser['name']), strtolower($browser)) !== false ){
-                    if ( !empty($version) ){
-                        if ( nebula_compare_operator($actual_version[0], $version_parts[0], $comparison) ){ //Major version comparison
-                            if ( !empty($version_parts[1]) ){ //If minor version exists and is not 0
-                                if ( nebula_compare_operator($actual_version[1], $version_parts[1], $comparison) ){ //Minor version comparison
-                                    return true;
-                                } else {
-                                    return false;
-                                }
-                            } else {
-                                return true;
-                            }
-                        }
-                    } else {
-                        return true;
-                    }
-                }
-            }
-
-            return false;
-        }
-
-        //Check to see if the rendering engine matches a passed parameter.
-        public function is_engine($engine=null){
-            $override = apply_filters('pre_nebula_is_engine', false, $engine);
-            if ( $override !== false ){return $override;}
-
-            if ( nebula_option('device_detection') ){
-                if ( empty($engine) ){
-                    trigger_error('nebula_is_engine requires a parameter of requested engine.');
-                    return false;
-                }
-
-                switch ( strtolower($engine) ){
-                    case 'ie':
-                    case 'internet explorer':
-                        $engine = 'trident';
-                        break;
-                    case 'web kit':
-                        $engine = 'webkit';
-                        break;
-                }
-
-                $actual_engine = $GLOBALS["device_detect"]->getClient();
-                if ( strpos(strtolower($actual_browser['engine']), strtolower($engine)) !== false ){
-                    return true;
-                }
-            }
-
-            return false;
-        }
-
-        //Check for bot/crawler traffic
-        //UA lookup: http://www.useragentstring.com/pages/Crawlerlist/
-        public function is_bot(){
-            $override = apply_filters('pre_nebula_is_bot', false);
-            if ( $override !== false ){return $override;}
-
-            if ( nebula_option('device_detection') ){
-                if ( $GLOBALS["device_detect"]->isBot() ){
-                    return true;
-                }
-            }
-
-            $bots = array('bot', 'crawl', 'spider', 'feed', 'slurp', 'tracker', 'http');
-            foreach( $bots as $bot ){
-                if ( strpos(strtolower($_SERVER['HTTP_USER_AGENT']), $bot) !== false ){
-                    return true;
-                    break;
-                }
-            }
-
-            return false;
         }
 
     }
